+++ src/main/java/com/daw/face/repository/model/User.java
@@ -1,8 +1,8 @@
 package com.daw.face.repository.model;
 
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 import lombok.ToString;
 
 import javax.persistence.Column;
@@ -11,6 +11,7 @@ import javax.persistence.Id;
 import javax.persistence.Table;
 
 @AllArgsConstructor
+@NoArgsConstructor
 @Data
 @ToString
 @Table(name = "\"user\"")
@@ -19,7 +20,6 @@ public class User {
     @Id
     @Column(name = "id_user")
     private String userId;
-
     @Column(name = "last_name")
     private String Name;
     @Column(name = "first_name")
@@ -30,8 +30,4 @@ public class User {
     private String password;
     @Column(name = "picture")
     private String pictureUrl;
-
-    public User() {
-
-    }
 }
